--- a/source/aasm/attrs.lisp
+++ b/source/aasm/attrs.lisp
@@ -132,7 +132,10 @@ out <- x ^ y (if integer)
 ```
 out <- move(x, y)
 where move(x, y) is x = y
-```")
+```
+- _jit_dont_render_me[boolean] (TODO)
+"
+	 :slots ((_jit_dont_render_me)))
 
 (defnode (:BinaryOps :MAX) (BinaryOps)
 	 "Computes the maximum value of two tensors in read, writing the result to the first write.