--- a/source/aasm/caten.aasm.asd
+++ b/source/aasm/caten.aasm.asd
@@ -7,6 +7,7 @@
   :components
   ((:file "package")
    (:file "helpers")
+   (:file "attrs")
    (:file "ctx")
    (:file "tensor-ir")
    (:file "ops")