--- a/source/aasm/helpers.lisp
+++ b/source/aasm/helpers.lisp
@@ -27,10 +27,11 @@
 			append (list (format nil "~a" n) ", ")))))
 
 (defun render-attrs (node &key (except-for nil))
-  (if (node-attrs node)	      
-      (with-output-to-string (out)
-	(format out " where")
-	(dolist (k (getattrs node))
-	  (when (and k (null (find k except-for)))
-	    (format out " :~(~a~)=~a" k (getattr node k)))))
-      ""))
+  (let ((attrs (dump-into-list (node-attr node) :allow-unbound nil)))
+    (if attrs
+	(with-output-to-string (out)
+	  (format out " where")
+	  (dolist (k (getattrs node))
+	    (when (and k (null (find k except-for)))
+	      (format out " :~(~a~)=~a" k (getattr node k)))))
+	"")))