- [x] Pass all tests w/ VM=1
- [x] Use placeholder instead of (car node-writes ...) ?
- [x] we do not want to initialize a numerous amount of CLOS classes
- [ ] fix for double-float in ccl
- [x] docgen
- [ ] Check they are dumpable and merge