@hikettei 2024-09-16
Initializing attributes 1000000 times, thanks to SBCL's optimization, is much faster than even making nodes.
```
CATEN/AASM> (time
	     (dotimes (i 1000000)
	       (make-node :A :A nil nil)))
Evaluation took:
  0.104 seconds of real time
  0.103185 seconds of total run time (0.100795 user, 0.002390 system)
  99.04% CPU
  288,025,856 bytes consed
  
NIL
CATEN/AASM> (time
	     (dotimes (i 1000000)
	       (make-instance 'NEG-Attr :_type_relay t)))
Evaluation took:
  0.013 seconds of real time
  0.013103 seconds of total run time (0.012579 user, 0.000524 system)
  100.00% CPU
  41 lambdas converted
  112,890,368 bytes consed
  
NIL
CATEN/AASM> (time
	     (dotimes (i 1000000)
	       (make-instance 'NEG-Attr :_type_relay t)))
Evaluation took:
  0.009 seconds of real time
  0.009780 seconds of total run time (0.009761 user, 0.000019 system)
  111.11% CPU
  111,973,680 bytes consed
  
NIL
CATEN/AASM> 
```