@hikettei 2024-09-16
No significant performance decrease were seen in main vs this PR
main:
JIT=1
```
* (time (caten (!rand `(n))))
Evaluation took:
  4.672 seconds of real time
  4.187063 seconds of total run time (4.103280 user, 0.083783 system)
  89.62% CPU
  48 forms interpreted
  231 lambdas converted
  176,127,920 bytes consed
```

This PR
```
CATEN> (time (caten (!rand `(n))))
Evaluation took:
  5.627 seconds of real time
  4.535812 seconds of total run time (4.094729 user, 0.441083 system)
  80.61% CPU
  48 forms interpreted
  231 lambdas converted
  182,998,464 bytes consed
```