@hikettei 2024-09-16
JIT=0
main:
```
* Evaluation took:
  2.872 seconds of real time
  2.729233 seconds of total run time (2.692141 user, 0.037092 system)
  95.02% CPU
  138,688,272 bytes consed
```

```
CATEN> (time (caten (!rand `(n))))
Evaluation took:
  2.822 seconds of real time
  2.728588 seconds of total run time (2.671705 user, 0.056883 system)
  96.70% CPU
  143,662,192 bytes consed
```